refactor: Move config.dart into app dir
